IMPORTANT: RSPEC AND DEVISE TOGETHER!
Never would I use HIGH CAPS, and especially not in a commit message,
would this not be important commit. This solves logging in as
superadmin, but same approach can and will be used for other
user levels to properly test all three test types that are influenced
by this, namely controllers and views via Devise Rspec helpers and
requests without Devise help, so we do manual authentication.

Signed-off-by: zeljko <zeljko@z-ware.fi>